Completely redone GKSession handling.
The session is now handled in the DataModel. Requests from views go into
the data model instance, states of the session are stored in core data
and a nice NSFetchedResultsController, using the FetchedResultsBackedTableViewController
handles the updates of the device lists.

The FirstViewController is now significantly lighter weight since it no longer
handles managing the session and connected peers. Having core data notify of changes
is fantastic. There are still a few connection use cases I need to regression test and then
fix but this is a great stopping point.