Refactor RJSessionManager and found a HUGE bug
1) I've refactored the session logic out of the data model. Now we
have a clean Session manager singleton and a data model singleton.

2) There was a bug where I was not passing in the predicate through a
convenience method and was therefore getting results that I did not expect.
I was seeing that the peer that was fetched was not the peer that I was
requesting. I'm now passing in the appopriate NSPredicate and getting the
appropriate device. Ugh.