A few changes to tighten up the experience
Now fetching by peerId, then device name, then create if not exists

Added a server image

Added a TextView for the header  on the FourthViewController to help explain what the view does

Added a button on the third view controller and a header if you are not in a party. This prevents a crash.

Changed the interface to the session so we get the currentServer a bit more cleanly.