Extract out the Music Specific message functionality from RJSessionManager
The problem I was having is that the RJSessionManager was specific to this
music use case. Now the RJSessionManager can be used for other things like the iPad
browsing experience or whatever. The DataModel should be interchangeable and the
message protocol is implemented in a subclass of the SessionManager.

Awesome.